fixed errors

+++ Task.java
@@ -1,11 +1,11 @@
 public abstract class Task {
     private String name;
     private String type;
-    private float startTime;
-    private float duration;
+    private double startTime;
+    private double duration;
     private int date;
     
-    public Task(String name, String type, float startTime, float duration, int date) {
+    public Task(String name, String type, double startTime, double duration, int date) {
         this.name = name;
         this.type = type;
         this.startTime = startTime;
